requirements.txt
opencv-python==4.9.0.80
numpy==1.23.5
torch==2.1.2+cu118
PyYAML==6.0.1
Pillow==10.1.0
datasets==2.16.1
timm==0.9.16
torchvision==0.16.2+cu118
safetensors==0.4.2
matplotlib==3.8.2
tabulate==0.9.0
sympy==1.12
overrides==7.4.0
transformers==4.37.1
onnx==1.13.0
typing_extensions==4.9.0
onnxruntime==1.13.1
dill==0.3.7
tensorboard==2.15.1
packaging==23.2
setuptools==68.2.0
tqdm==4.66.1
scipy==1.11.4
numba==0.57.1
psutil==5.9.7
colorama==0.4.6
urllib3==2.0.7
fsspec==2023.10.0
astroid==3.1.0
ipython==8.12.3
traitlets==5.14.1
SQLAlchemy==2.0.30
pyarrow==15.0.0
requests==2.31.0
multiprocess==0.70.15
filelock==3.13.1
aiohttp==3.9.1
pandas==2.1.4
jax==0.4.25
xxhash==3.4.1
arrow==1.3.0
torchaudio==2.1.2+cu118
soundfile==0.12.1
av==11.0.0
defusedxml==0.7.1
scikit-learn==1.3.2
openai==1.11.1
regex==2023.12.25
tokenizers==0.15.1
beautifulsoup4==4.12.3
librosa==0.10.2
h5py==3.10.0
msgpack==1.0.7
insightface==0.7.3
PyQt5==5.15.10
python-dotenv==1.0.1
scikit-image==0.22.0
mxnet==1.8.0